Weather-Project/requirements.txt
requests==2.31.0
bcrypt==4.1.1
Pillow==10.1.0
